+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Fixed
+- Page swipe bypasses block-page-navigation-on-new-constraint-error feature.
+
 [4.35.2] - 2017-06-13
 ---------------------
 ##### Fixed